Updating the src to include the menu item
