+++ src/blocks.js
@@ -1,0 +1,2 @@
+import './mega-menu/index';
+import './menu-item/index';
