+++ src/menu-item/edit.js
@@ -1,0 +1,42 @@
+import { SelectControl } from '@wordpress/components';
+
+/**
+ * Output our menu selector.
+ *
+ * @return {WPElement} Element to render.
+ */
+export default function Edit( props ) {
+	const { useSelect } = wp.data;
+
+	// Get the wp_template_parts
+	const templateParts = useSelect( ( select ) => {
+		return select( 'core' ).getEntityRecords( 'postType', 'wp_template_part' )
+	} )
+
+	let html = '';
+
+	if ( null === templateParts ) {
+		html = ( <div>Loading</div> );
+	} else {
+
+		let options = [];
+		options.push( { value: '', label: 'None' } );
+		templateParts.forEach( ( part ) => {
+			if ( -1 < part.slug.search('mega-menu')  ) {
+				options.push( { value : part.slug, label : part.title.rendered } )
+			}
+		});
+
+		html = (
+			<SelectControl
+				label={ 'Select a mega menu:' }
+				value={ props.attributes.menu }
+				onChange={ ( menu ) => {
+					props.setAttributes( { menu } );
+				} }
+				options={ options }
+			/>
+		);
+	}
+	return html;
+}
