+++ src/menu-item/index.js
@@ -1,0 +1,13 @@
+import { registerBlockType } from '@wordpress/blocks';
+
+import Edit from './edit';
+import Save from './save';
+import metadata from './block.json';
+
+registerBlockType( 
+	metadata.name, {
+		title: metadata.title,
+		edit: Edit,
+		save: Save,
+	}
+);
