--- src2/link-ui.js
@@ -1,208 +1,0 @@
-import { createElement } from "@wordpress/element";
-
-/**
- * WordPress dependencies
- */
-import { __unstableStripHTML as stripHTML } from '@wordpress/dom';
-import { Popover, Button } from '@wordpress/components';
-import { __, sprintf } from '@wordpress/i18n';
-import { __experimentalLinkControl as LinkControl, BlockIcon, store as blockEditorStore } from '@wordpress/block-editor';
-import { createInterpolateElement } from '@wordpress/element';
-import { store as coreStore } from '@wordpress/core-data';
-import { decodeEntities } from '@wordpress/html-entities';
-import { switchToBlockType } from '@wordpress/blocks';
-import { useSelect, useDispatch } from '@wordpress/data';
-/**
- * Given the Link block's type attribute, return the query params to give to
- * /wp/v2/search.
- *
- * @param {string} type Link block's type attribute.
- * @param {string} kind Link block's entity of kind (post-type|taxonomy)
- * @return {{ type?: string, subtype?: string }} Search query params.
- */
-
-export function getSuggestionsQuery(type, kind) {
-  switch (type) {
-    case 'post':
-    case 'page':
-      return {
-        type: 'post',
-        subtype: type
-      };
-
-    case 'category':
-      return {
-        type: 'term',
-        subtype: 'category'
-      };
-
-    case 'tag':
-      return {
-        type: 'term',
-        subtype: 'post_tag'
-      };
-
-    case 'post_format':
-      return {
-        type: 'post-format'
-      };
-
-    default:
-      if (kind === 'taxonomy') {
-        return {
-          type: 'term',
-          subtype: type
-        };
-      }
-
-      if (kind === 'post-type') {
-        return {
-          type: 'post',
-          subtype: type
-        };
-      }
-
-      return {};
-  }
-}
-/**
- * Add transforms to Link Control
- *
- * @param {Object} props          Component props.
- * @param {string} props.clientId Block client ID.
- */
-
-function LinkControlTransforms(_ref) {
-  let {
-    clientId
-  } = _ref;
-  const {
-    getBlock,
-    blockTransforms
-  } = useSelect(select => {
-    const {
-      getBlock: _getBlock,
-      getBlockRootClientId,
-      getBlockTransformItems
-    } = select(blockEditorStore);
-    return {
-      getBlock: _getBlock,
-      blockTransforms: getBlockTransformItems(_getBlock(clientId), getBlockRootClientId(clientId))
-    };
-  }, [clientId]);
-  const {
-    replaceBlock
-  } = useDispatch(blockEditorStore);
-  const featuredBlocks = ['core/page-list', 'core/site-logo', 'core/social-links', 'core/search'];
-  const transforms = blockTransforms.filter(item => {
-    return featuredBlocks.includes(item.name);
-  });
-
-  if (!(transforms !== null && transforms !== void 0 && transforms.length)) {
-    return null;
-  }
-
-  if (!clientId) {
-    return null;
-  }
-
-  return createElement("div", {
-    className: "link-control-transform"
-  }, createElement("h3", {
-    className: "link-control-transform__subheading"
-  }, __('Transform')), createElement("div", {
-    className: "link-control-transform__items"
-  }, transforms.map((item, index) => {
-    return createElement(Button, {
-      key: `transform-${index}`,
-      onClick: () => replaceBlock(clientId, switchToBlockType(getBlock(clientId), item.name)),
-      className: "link-control-transform__item"
-    }, createElement(BlockIcon, {
-      icon: item.icon
-    }), item.title);
-  })));
-}
-
-export function LinkUI(props) {
-  const {
-    saveEntityRecord
-  } = useDispatch(coreStore);
-
-  async function handleCreate(pageTitle) {
-    const postType = props.link.type || 'page';
-    const page = await saveEntityRecord('postType', postType, {
-      title: pageTitle,
-      status: 'draft'
-    });
-    return {
-      id: page.id,
-      type: postType,
-      // Make `title` property consistent with that in `fetchLinkSuggestions` where the `rendered` title (containing HTML entities)
-      // is also being decoded. By being consistent in both locations we avoid having to branch in the rendering output code.
-      // Ideally in the future we will update both APIs to utilise the "raw" form of the title which is better suited to edit contexts.
-      // e.g.
-      // - title.raw = "Yes & No"
-      // - title.rendered = "Yes &#038; No"
-      // - decodeEntities( title.rendered ) = "Yes & No"
-      // See:
-      // - https://github.com/WordPress/gutenberg/pull/41063
-      // - https://github.com/WordPress/gutenberg/blob/a1e1fdc0e6278457e9f4fc0b31ac6d2095f5450b/packages/core-data/src/fetch/__experimental-fetch-link-suggestions.js#L212-L218
-      title: decodeEntities(page.title.rendered),
-      url: page.link,
-      kind: 'post-type'
-    };
-  }
-
-  const {
-    label,
-    url,
-    opensInNewTab,
-    type,
-    kind
-  } = props.link;
-  const link = {
-    url,
-    opensInNewTab,
-    title: label && stripHTML(label)
-  };
-  return createElement(Popover, {
-    placement: "bottom",
-    onClose: props.onClose,
-    anchor: props.anchor,
-    __unstableSlotName: '__unstable-block-tools-after',
-    shift: true
-  }, createElement(LinkControl, {
-    hasTextControl: true,
-    hasRichPreviews: true,
-    className: props.className,
-    value: link,
-    showInitialSuggestions: true,
-    withCreateSuggestion: props.hasCreateSuggestion,
-    createSuggestion: handleCreate,
-    createSuggestionButtonText: searchTerm => {
-      let format;
-
-      if (type === 'post') {
-        /* translators: %s: search term. */
-        format = __('Create draft post: <mark>%s</mark>');
-      } else {
-        /* translators: %s: search term. */
-        format = __('Create draft page: <mark>%s</mark>');
-      }
-
-      return createInterpolateElement(sprintf(format, searchTerm), {
-        mark: createElement("mark", null)
-      });
-    },
-    noDirectEntry: !!type,
-    noURLSuggestion: !!type,
-    suggestionsQuery: getSuggestionsQuery(type, kind),
-    onChange: props.onChange,
-    onRemove: props.onRemove,
-    onCancel: props.onCancel,
-    renderControlBottom: !url ? () => createElement(LinkControlTransforms, {
-      clientId: props.clientId
-    }) : null
-  }));
-}
-//# sourceMappingURL=link-ui.js.map
